requirements.txt
selenium
requests
PyYAML
